Reorganize a bit the test files
